requirements.txt
aiofiles==0.6.0
asgiref==3.4.1
bcrypt==3.2.0
blis==0.7.4
catalogue==2.0.4
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.3
click==7.1.2
cryptography==3.4.7
cymem==2.0.5
Deprecated==1.2.12
ecdsa==0.17.0
fastapi==0.66.1
googleapis-common-protos==1.53.0
grpcio==1.39.0
h11==0.12.0
idna==3.2
Jinja2==2.11.2
joblib==1.0.1
MarkupSafe==2.0.1
murmurhash==1.0.5
nltk==3.6.2
numpy==1.21.1
opentelemetry-api==1.4.0
opentelemetry-exporter-jaeger==1.4.0
opentelemetry-exporter-jaeger-proto-grpc==1.4.0
opentelemetry-exporter-jaeger-thrift==1.4.0
opentelemetry-instrumentation==0.23b0
opentelemetry-instrumentation-asgi==0.23b0
opentelemetry-instrumentation-fastapi==0.23b0
opentelemetry-sdk==1.4.0
opentelemetry-semantic-conventions==0.23b0
opentelemetry-util-http==0.23b0
packaging==21.0
passlib==1.7.4
pathy==0.6.0
preshed==3.0.5
protobuf==3.17.3
pyasn1==0.4.8
pycparser==2.20
pydantic==1.8.2
pyparsing==2.4.7
python-jose==3.3.0
python-multipart==0.0.5
regex==2021.7.6
requests==2.26.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.0
six==1.16.0
sklearn==0.0
smart-open==5.1.0
spacy==3.1.1
spacy-legacy==3.0.8
spacytextblob==3.0.1
srsly==2.4.1
starlette==0.14.2
textblob==0.15.3
thinc==8.0.8
threadpoolctl==2.2.0
thrift==0.13.0
tqdm==4.61.2
typer==0.3.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.12.3
wasabi==0.8.2
wrapt==1.12.1
